+++ core-model/Cargo.toml
@@ -2,7 +2,7 @@
 authors = ["Terkwood <38859656+Terkwood@users.noreply.github.com>"]
 edition = "2018"
 name = "core-model"
-version = "0.2.0"
+version = "0.2.1"
 
 [dependencies]
 serde = "1.0.116"
