+++ micro-color-chooser/Cargo.toml
@@ -6,17 +6,16 @@ name = "micro-color-chooser"
 repository = "https://github.com/Terkwood/BUGOUT"
 version = "0.1.1"
 
+[dev-dependencies]
+crossbeam-channel = "0.4.4"
+
 [dependencies]
 bincode = "1.3.1"
-core-model = {git = "https://github.com/Terkwood/BUGOUT", rev = "af087ee"}
+color-model = {path = "color-model"}
+core-model = {git = "https://github.com/Terkwood/BUGOUT", branch = "fix/color-choose-model"}
 env_logger = "0.7.1"
 log = "0.4.11"
 rand = "0.7.3"
 redis = {version = "0.17.0", features = ["r2d2"]}
 redis_streams = {git = "https://github.com/Terkwood/BUGOUT", version = "0.3.1", rev = "0e2fdb7"}
-serde = "1.0.116"
-serde_derive = "1.0.116"
 uuid = {version = "0.8.1", features = ["serde", "v4"]}
-
-[dev-dependencies]
-crossbeam-channel = "0.4.4"
