+++ micro-color-chooser/color-model/src/lib.rs
@@ -1,7 +1,31 @@
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
+pub mod api;
+
+use core_model::*;
+use serde_derive::{Deserialize, Serialize};
+
+#[derive(Copy, Clone, Debug, Deserialize, Serialize, PartialEq)]
+pub enum ColorPref {
+    Black,
+    White,
+    Any,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+pub struct SessionColorPref {
+    pub session_id: SessionId,
+    pub color_pref: ColorPref,
+    pub client_id: ClientId,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize, PartialEq)]
+pub enum GameColorPref {
+    NotReady,
+    Partial {
+        game_id: GameId,
+        pref: SessionColorPref,
+    },
+    Complete {
+        game_id: GameId,
+        prefs: (SessionColorPref, SessionColorPref),
+    },
 }
