+++ micro-color-chooser/src/repo/game_ready.rs
@@ -1,5 +1,5 @@
 use super::*;
-use crate::api::GameReady;
+use color_model::api::GameReady;
 use core_model::SessionId;
 use redis::Commands;
